// src/app/model/request/student-parent-profile-request.ts
export interface StudentParentProfileRequest {
  parentId: number;
  fatherName: string;
  fatherOccupation: string;
  fatherPhoneNumber: string;
  motherName: string;
  motherOccupation: string;
  motherPhoneNumber: string;
  email: string;
  doorNum: string;
  street: string;
  addrLine: string;
  city: string;
  state: string;
  pincode: number;
}